Исправлены недочеты
