+++ src/components/burger-constructor/burger-constructor.module.css
@@ -6,7 +6,12 @@
 .constructor-container {
     display: grid;
     grid-gap: 18px;
-    max-height: 500px;
+    max-height: 300px;
+}
+
+.constructor-wrapper {
+    display: grid;
+    grid-gap: 16px;
 }
 
 .list {
