+++ src/components/burger-ingredients/burger-ingredients.module.css
@@ -21,3 +21,7 @@
     overflow: auto;
     max-height: 500px;
 }
+
+.tabs {
+    display: flex;
+}
